Revert "Merge pull request #17072 from yashykt/interceptor_cancellation"
This reverts commit 193b4b57ec8ec1509a91069f8cdd1ba1917b2d71, reversing
changes made to 385fde64fc2e3b0241229e4a018ea0989a2a00bc.
