+++ include/grpcpp/impl/codegen/client_context.h
@@ -426,8 +426,6 @@ class ClientContext {
 
   grpc::string authority() { return authority_; }
 
-  void SendCancelToInterceptors();
-
   bool initial_metadata_received_;
   bool wait_for_ready_;
   bool wait_for_ready_explicitly_set_;
