+++ src/cpp/client/channel_cc.cc
@@ -147,14 +147,10 @@ internal::Call Channel::CreateCallInternal(const internal::RpcMethod& method,
     }
   }
   grpc_census_call_set_context(c_call, context->census_context());
-
-  // ClientRpcInfo should be set before call because set_call also checks
-  // whether the call has been cancelled, and if the call was cancelled, we
-  // should notify the interceptors too/
-  auto* info = context->set_client_rpc_info(
-      method.name(), this, interceptor_creators_, interceptor_pos);
   context->set_call(c_call, shared_from_this());
 
+  auto* info = context->set_client_rpc_info(
+      method.name(), this, interceptor_creators_, interceptor_pos);
   return internal::Call(c_call, this, cq, info);
 }
