+++ test/cpp/end2end/BUILD
@@ -38,7 +38,6 @@ grpc_cc_library(
 grpc_cc_library(
     name = "interceptors_util",
     testonly = True,
-    srcs = ["interceptors_util.cc"],
     hdrs = ["interceptors_util.h"],
     external_deps = [
         "gtest",
@@ -159,7 +158,6 @@ grpc_cc_library(
         "gtest",
     ],
     deps = [
-        ":interceptors_util",
         ":test_service_impl",
         "//:gpr",
         "//:grpc",
